pyproject.toml
[tool.poetry]
name = "sigmakokicommander"
version = "0.1.0"
description = ""
authors = ["PlusF <rkaneda896@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.6.8"
pyserial = ">=3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
